CI: tidy-check'in gercek basarisizlik nedenini duzelt + kalan clang-tidy uyarilarini gider
tidy-check job'u compile_commands.json'i indiriyordu ama
GenerateExportHeader'in urettigi generated_include/*/export.h
dosyalarini indirmiyordu -- bu yuzden export.h iceren HER dosya
"dosya bulunamadi" derleyici hatasiyla bastan basarisiz oluyordu
(informative/continue-on-error oldugu icin CI'i kirmiyordu ama
tidy-check pratikte hicbir seyi analiz edemiyordu). Artifact'e
generated_include de eklendi.

Bu duzeltme sayesinde gercekten calisan clang-tidy, ChunkGrid.h ve
TerrainDeformSystem.cpp'de int->float bugprone-narrowing-conversions
uyarilari buldu (chunk koordinati * chunk boyutu carpimlari) --
explicit static_cast ile giderildi. src altinda artik sifir
clang-tidy uyarisi/hatasi (lokal olarak dogrulandi).

+++ src/terrain_deform/TerrainDeformSystem.cpp
@@ -6,7 +6,7 @@ namespace alazforge {
 
 TerrainDeformSystem::TerrainDeformSystem(const Config& inConfig)
     : config(inConfig),
-      cellsPerSide(static_cast<int>(inConfig.chunkSize / inConfig.cellSize)),
+      cellsPerSide(static_cast<int>(static_cast<float>(inConfig.chunkSize) / inConfig.cellSize)),
       grid(inConfig.mapSize, inConfig.chunkSize),
       stream(grid, inConfig.loadRadius, inConfig.savePath) {}
